+++ src/python/pants/backend/scala/compile/scalac.py
@@ -26,6 +26,7 @@ from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import CoarsenedTarget, CoarsenedTargets, FieldSet, SourcesField, Targets
 from pants.engine.unions import UnionRule
 from pants.jvm.compile import CompiledClassfiles, CompileResult, FallibleCompiledClassfiles
+from pants.jvm.compile import rules as jvm_compile_rules
 from pants.jvm.jdk_rules import JdkSetup
 from pants.jvm.resolve.coursier_fetch import (
     ArtifactRequirements,
@@ -225,15 +226,6 @@ async def compile_scala_source(
     )
 
 
-@rule
-def required_classfiles(fallible_result: FallibleCompiledClassfiles) -> CompiledClassfiles:
-    if fallible_result.result == CompileResult.SUCCEEDED:
-        assert fallible_result.output
-        return fallible_result.output
-    # NB: The compile outputs will already have been streamed as FallibleCompiledClassfiles finish.
-    raise Exception("Compile failed.")
-
-
 @rule(desc="Check compilation for Scala", level=LogLevel.DEBUG)
 async def scalac_check(request: ScalacCheckRequest) -> CheckResults:
     coarsened_targets = await Get(
@@ -265,5 +257,6 @@ async def scalac_check(request: ScalacCheckRequest) -> CheckResults:
 def rules():
     return [
         *collect_rules(),
+        *jvm_compile_rules(),
         UnionRule(CheckRequest, ScalacCheckRequest),
     ]
